requirements.txt
flask
pyjwt
bcrypt
email-validator
flask-cors
gunicorn
python-dotenv
flask-limiter
marshmallow
Pillow